make EurekaClient available before smart lifecycle phase 0
